Moved angle processing to it's own subsystem so commands on the shooter and the shooter angle could run in parallel
